Move socket based operation into new file
Add new file for socket opeartions, for common use
across different OS.
